+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add button to switch between list and edit view
+
+### Updated
+
+- Update layout to display list or edit view only
+- Display list view by default if id param present
+- Display edit view by default if no id param specified
+
 ## [1.2.1] - 2025-08-18
 
 ### Updated